2023-6-9 初次提交版

+++ routers/api/v2/deletefile.go
@@ -1,10 +1,10 @@
 package v2
 
 import (
+	"fmt"
 	"pan/pkg/app"
 	"pan/pkg/errcode"
 	"pan/utils"
-	"strconv"
 
 	"github.com/gin-gonic/gin"
 )
@@ -14,20 +14,23 @@ func DeleteFile() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 
 		response := app.NewRespponse(ctx)
-		userid := ctx.GetUint("UserID")
-		filepath := ctx.PostForm("filepath")
 
-		path := "storage" + "/" + strconv.Itoa(int(userid)) + "/" + filepath
+		userid := ctx.GetUint("UserID")
+
+		floderpath := ctx.DefaultPostForm("floderpath", "")
+		filename := ctx.DefaultPostForm("filename", "")
+		if filename == "" {
+			response.ToErrorResponse(errcode.CanNotFindFile.WithDetails("没有指定文件"))
+			return
+		}
+
+		path := fmt.Sprintf("storage/%d/%s/%s", userid, floderpath, filename)
 		err := utils.RemoveFile(path)
 		if err != nil {
 			response.ToErrorResponse(errcode.DeleteFileError)
 			return
 		}
 
-		data := gin.H{
-			"code":    0,
-			"message": "Delete success!",
-		}
-		response.ToResponse(data)
+		response.ToSuccessResponse("Delete success!")
 	}
 }
